+++ lib/nostr_dart.dart
@@ -2,5 +2,7 @@ library nostr;
 
 export 'src/nostr.dart';
 export 'src/keys.dart';
-export 'src/model/event.dart';
-export 'src/model/contact_list.dart' show Contact, ContactList;
+export 'src/event.dart';
+export 'src/contact_list.dart' show Contact, ContactList;
+export 'src/relay.dart';
+export 'src/relay_info.dart';