+++ lib/src/relay.dart
@@ -1,37 +1,70 @@
 import 'dart:convert';
 import 'dart:async';
-import 'dart:io';
 import 'dart:collection';
 import 'dart:developer';
 import 'package:web_socket_channel/io.dart';
 import 'package:queue/queue.dart' as job_queue;
+import 'package:http/http.dart' as http;
+import 'relay_info.dart';
+import 'relay_info_provider.dart';
 
 enum WriteAccess { readOnly, writeOnly, readWrite }
 
+/// Represents an individual relay and manages communications to it.
 class Relay {
-  final String _url;
-  final WriteAccess access;
+  /// Relay's URL
+  final String url;
+
+  /// Relay read/write access configuration
+  WriteAccess access;
+
+  /// [NIP-11](https://github.com/nostr-protocol/nips/blob/master/11.md) relay information
+  late RelayInfo info;
+
   late WebSocketHandler _ws;
   final _jobRunner = job_queue.Queue();
   final Queue<Completer<void>> _pendingResponses = Queue();
   Function? _listener;
+  Function(String)? onError;
+  final RelayInfoProvider _relayInfoProvider;
 
-  Relay(String url, {Function? onDone, this.access = WriteAccess.readOnly})
-      : _url = url {
-    _ws = WebSocketHandler(_url, onDone: onDone);
+  /// Creates a `Relay` object.
+  ///
+  /// [url] is the URL of the relay.
+  /// [access] specifies read/write access for the added relay and may be set to
+  /// either `WriteAccess.readOnly`, `WriteAccess.writeOnly`, or
+  /// `WriteAccess.readWrite`. If not provided the default is read-only.
+  /// [relayInfoProvider] is used for testing and not intended for normal use.
+  Relay(this.url,
+      {this.access = WriteAccess.readOnly,
+      this.onError,
+      RelayInfoProvider? relayInfoProvider})
+      : _relayInfoProvider =
+            relayInfoProvider ?? RelayInfoProvider(http: http.Client()) {
+    _ws = WebSocketHandler(url);
     _ws.addListener(_onData);
   }
 
-  String get url => _url;
-
-  bool get isDisconnected => _ws.isDisconnected;
+  /// Relay connection status. `true` if disconnected.
+  bool get isConnected => _ws.isConnected;
 
   void listen(Function? callback) {
     _listener = callback;
   }
 
   Future<bool> connect() async {
-    bool result = await _ws.connect();
+    bool result = false;
+    try {
+      info = await _relayInfoProvider.get(url);
+
+      // Relay must support NIP-15 and NIP-20
+      if (info.nips.contains(15) && info.nips.contains(20)) {
+        await _ws.connect();
+        result = true;
+      }
+    } catch (e) {
+      log(e.toString());
+    }
     return result;
   }
 
@@ -45,19 +78,23 @@ class Relay {
 
   Future<void> _send(String message) async {
     final completer = Completer();
-    completer.future.timeout(
-      const Duration(seconds: 5),
-      onTimeout: () {
-        throw TimeoutException("No response from $_url");
-      },
-    ).catchError((err) {
-      log(err.toString());
+    completer.future.timeout(const Duration(seconds: 5), onTimeout: () {
+      log("No response from $url");
       _pendingResponses.removeFirst().complete();
       disconnect();
     });
     _pendingResponses.add(completer);
-    log("Sending to $_url");
-    _ws.send(message);
+
+    try {
+      _ws.send(message);
+    } catch (e) {
+      _pendingResponses.removeFirst().complete();
+      disconnect();
+      if (onError != null) {
+        onError!(url);
+      }
+    }
+
     await completer.future;
   }
 
@@ -69,10 +106,10 @@ class Relay {
       }
     }
     if (json[0] == 'NOTICE') {
-      log("$_url: ${json.toString()}");
+      log("$url: ${json.toString()}");
     }
     if (_listener != null) {
-      json.add(_url);
+      json.add(url);
       _listener!(json);
     }
   }
@@ -81,50 +118,40 @@ class Relay {
 class WebSocketHandler {
   final String _url;
   IOWebSocketChannel? _channel;
-  bool _disconnected = true;
+  bool _connected = false;
   final List<Function> _listeners = [];
-  final Function? _onDone;
 
-  WebSocketHandler(this._url, {Function? onDone}) : _onDone = onDone;
+  WebSocketHandler(this._url);
 
-  bool get isDisconnected => _disconnected;
+  bool get isConnected => _connected;
 
-  Future<bool> connect() async {
+  Future<void> connect() async {
     reset();
-    try {
-      await WebSocket.connect(_url)
-          .timeout(const Duration(seconds: 10))
-          .then((ws) {
-        _channel = IOWebSocketChannel(ws);
-        _disconnected = false;
-        _channel!.stream.listen(_onReceiveMessage, onError: (error) {
-          log("Websocket stream error: $_url");
-        }, onDone: () {
-          _disconnected = true;
-          if (_onDone != null) {
-            _onDone!(_url);
-          }
-        });
-      });
-      return true;
-    } catch (err) {
-      log(err.toString());
-      return false;
-    }
+    log("Connecting to $_url");
+    _channel = IOWebSocketChannel.connect(_url,
+        connectTimeout: const Duration(seconds: 10));
+    _connected = true;
+    _channel!.stream.listen(_onReceiveMessage, onError: (error) async {
+      log("Websocket stream error: $_url");
+      await reset();
+    }, onDone: () {
+      log("Websocket stream closed by remote: $_url");
+      _connected = false;
+    });
   }
 
   Future<void> reset() async {
-    if (_channel != null) {
+    if (_connected) {
       await _channel!.sink.close();
-      _disconnected = true;
+      _connected = false;
     }
   }
 
   void send(String message) async {
-    // if (_disconnected) {
-    //   await connect();
-    // }
-    if (_channel != null && !_disconnected) {
+    if (!_connected) {
+      await connect();
+    }
+    if (_connected) {
       _channel!.sink.add(message);
     }
   }
@@ -138,7 +165,6 @@ class WebSocketHandler {
   }
 
   _onReceiveMessage(message) {
-    _disconnected = false;
     for (Function callback in _listeners) {
       callback(message);
     }